chore(test): move e2e specs to their corresponding e2e-app folders
