+++ angular.json
@@ -192,13 +192,13 @@
       }
     },
     "e2e": {
-      "root": "e2e-app/e2e",
+      "root": "e2e-app/src",
       "projectType": "application",
       "architect": {
         "e2e": {
           "builder": "@angular-devkit/build-angular:protractor",
           "options": {
-            "protractorConfig": "e2e-app/e2e/protractor.conf.js",
+            "protractorConfig": "e2e-app/protractor.conf.js",
             "devServerTarget": "e2e-app:serve"
           },
           "configurations": {
@@ -213,7 +213,7 @@
         "lint": {
           "builder": "@angular-devkit/build-angular:tslint",
           "options": {
-            "tsConfig": "e2e-app/e2e/tsconfig.json",
+            "tsConfig": "e2e-app/tsconfig.spec.json",
             "exclude": [
               "**/node_modules/**"
             ]
