+++ e2e-app/src/app/datepicker/focus/datepicker-focus.po.ts
@@ -1,5 +1,5 @@
 import {$} from 'protractor';
-import {DatepickerPage} from './datepicker.po';
+import {DatepickerPage} from './../datepicker.po';
 
 export class DatepickerFocusPage extends DatepickerPage {
   async preSelectDate() { await $('#selectDate').click(); }
